feat(pagination): refactor pagination into low-level
