+++ packages/clay-pagination/src/Ellipsis.tsx
