+++ packages/clay-pagination/src/index.tsx
@@ -6,7 +6,8 @@
 
 import ClayPagination from './Pagination';
 import {ClayPaginationWithBar} from './PaginationWithBar';
+import {ClayPaginationWithBasicItems} from './PaginationWithBasicItems';
 
-export {ClayPaginationWithBar};
+export {ClayPaginationWithBar, ClayPaginationWithBasicItems};
 
 export default ClayPagination;
